gen_rust_project: Get files from Bazel directly
When I run gen_rust_project with
`--bazel-aquery-flags=--platforms=//:myplatform`, the tool ends up reading a
crate_spec file whose output refers to the host triple instead of the desired
target triple. I don't know why this is happening, but I suspect the
relationship between the paths in the aquery output and the file the tool
constructs a path to and reads is not quite right wrt Bazel's model.

These crate_spec files are written via `ctx.actions.write()`, so this change
passes `--include_file_write_contents` to the aquery invocation and reads the
file contents from the `actions[*].file_contents` field of the result.

Using this method instead of locating and reading the crate_spec file gives me
the results I expect.